attestation_report_fetcher, snp_device: Split Linux specific code into _linux, add _windows using amdsnppspapi.dll
The dll itself needs to be fetched via some other means, and placed in the
container's System32.
