+++ pkg/attest/snp_device_linux.go
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+//go:build linux
+
 package attest
 
 import (
