+++ pkg/attest/snp_device_windows.go
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+//go:build windows
+
+package attest
+
+func IsSNPVM() bool {
+	_, err := NewAttestationReportFetcher()
+	return err == nil
+}
